[REF] product, *: save template specific attribute values on variants
* account, hr_expense, mrp, sale, stock_account, website_sale,
	website_sale_comparison, website_sale_wishlist

Before this commit, the `product.attribute.value` were stored on the variants.
This required filtering of attribute lines to find the appropriate matching
`product.template.attribute.value` that were used in most of the business code.

This also prevented to have multiple attribute lines for the same attribute,
which is something we want in the future, to handle use cases such as grape
varieties for wine products (more generally having multiple lines for the same
attribute with only one value on them), or such as having multiple colors on a
product without having to duplicate the color attribute.

Now the combination of `product.template.attribute.value` is directly stored on
the variant.

Also add combination_indices on product, which allows to quickly find a variant
matching a combination, and to add an easy sql constraints to ensure active
combination uniqueness.