[FIX] website, *: fix wishlist animation
* = website_sale, website_sale_wishlist

The wishlist animation was not handling the floating navbar and always
refering to the fixed one.
Now the animation will target the right navbar if we scroll down.

When the navbar is duplicated, the ids of the tags were kept wich is
wrong and was producoing bugs such as wishlist button not showing up
on the floating navbar on the first product added to the wishlist and
the first issue. Now the id is removed from the clone and never used
in JS code. It's only purpose is for the xpath of the wishlist button.

When a product is added to the cart from the wishlist and the floating
navbar is displayed, the product is hidden before the animation is
ompleted. This changes the height of the page and the animation goes to
the middle of the page instead of the button. This is now fixed too.

Task-2002122