[IMP] website_sale_slides: improve course purchase flow V2
Now we can have eCommerce and eLearning without automatically have the option to
sell courses.
The auto_install in the manifest was also blocking the possibility to uninstall the
module. When we tried to uncheck the option for selling course, it reinstalled it
just after we had saved the change in the settings.