[IMP] sale_management: add company on quotation template.
Purpose of the commit is to ensures that a user can choose
sale templates of his/her own company or the templates which
are common.

currently when the product sharing is disbled for multi company
and selecting the quotation template which contains the product of
another company it gives Error because adding a product from another
company into current quotation.

Also added company on demo data as the demo quotation template contains
the product of main company.

task-2025168
Closes: #34342