[FIX] pos_mercury: don't show amount due in payment method popup
This popup shows in the rare situation where somebody has 2 payment
methods that use Mercury. It allows them to choose which one to use
when swiping a card.

It would show the amount due in the title of the popup however this
isn't correct when a "pending swipe" payment line has been added.

Fix it by removing this amount, it's clear what will be paid from the
information on the payment screen.

opw-2006596